+++ app-runtime-management/installer/windows/build.ps1
@@ -1,6 +1,6 @@
 param(
   [string]$Configuration = 'Release',
-  [string]$Version = '2.0.4'
+  [string]$Version = '2.0.5'
 )
 
 $ErrorActionPreference = 'Stop'